+++ tests/index.js
@@ -56,6 +56,15 @@ class MockCompiler {
     }
 }
 
+
+class MockCompilerNoStat extends MockCompiler {
+    constructor (options = {}) {
+        super(options);
+
+        this.inputFileSystem.stat = (file, cb) => cb(undefined, undefined);
+    }
+}
+
 describe('apply function', () => {
     // Ideally we pass in patterns and confirm the resulting assets
     const run = (opts) => {
@@ -329,6 +338,57 @@ describe('apply function', () => {
             .catch(done);
         });
 
+        it('can transform target path of every file in glob', (done) => {
+            runEmit({
+                expectedAssetKeys: [
+                    '/some/path/(special-*file).txt.tst',
+                    '/some/path/binextension.bin.tst',
+                    '/some/path/file.txt.tst',
+                    '/some/path/file.txt.gz.tst',
+                    '/some/path/directoryfile.txt.tst',
+                    '/some/path/nestedfile.txt.tst',
+                    '/some/path/noextension.tst',
+                    '/some/path/hello.txt.tst'
+                ],
+                patterns: [{
+                    from: '**/*',
+                    transformPath: function(targetPath, absoluteFrom) {
+                        expect(absoluteFrom).to.have.string(HELPER_DIR);
+                        return '/some/path/' + path.basename(targetPath) + '.tst';
+                    }
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
+        it('can transform target path of every file in glob after applying template', (done) => {
+            runEmit({
+                expectedAssetKeys: [
+                    'transformed/[!]/hello-d41d8c.txt',
+                    'transformed/[special?directory]/directoryfile-22af64.txt',
+                    'transformed/[special?directory]/(special-*file)-0bd650.txt',
+                    'transformed/[special?directory]/nested/nestedfile-d41d8c.txt',
+                    'transformed/binextension-d41d8c.bin',
+                    'transformed/file-22af64.txt',
+                    'transformed/file.txt-5b311c.gz',
+                    'transformed/directory/directoryfile-22af64.txt',
+                    'transformed/directory/nested/nestedfile-d41d8c.txt',
+                    'transformed/noextension-d41d8c'
+                ],
+                patterns: [{
+                    from: '**/*',
+                    to: 'nested/[path][name]-[hash:6].[ext]',
+                    transformPath: function(targetPath, absoluteFrom) {
+                        expect(absoluteFrom).to.have.string(HELPER_DIR);
+                        return targetPath.replace('nested/', 'transformed/');
+                    }
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
         it('can use a glob to move multiple files in a different relative context to a non-root directory', (done) => {
             runEmit({
                 expectedAssetKeys: [
@@ -482,6 +542,37 @@ describe('apply function', () => {
             .then(done)
             .catch(done);
         });
+
+        it('adds the directory to the watch list when using glob', (done) => {
+            run({
+                patterns: [{
+                    from: 'directory/**/*'
+                }]
+            })
+            .then((compilation) => {
+                const absFrom = path.resolve(HELPER_DIR, 'directory');
+                const absFromNested = path.resolve(HELPER_DIR, 'directory', 'nested');
+                expect(compilation.contextDependencies).to.have.members([absFrom, absFromNested]);
+            })
+            .then(done)
+            .catch(done);
+        });
+
+        it('does not add the directory to the watch list when glob is a file', (done) => {
+            run({
+                patterns: [{
+                    from: {
+                        glob: 'directory/directoryfile.txt'
+                    }
+                }]
+            })
+            .then((compilation) => {
+                const absFrom = path.resolve(HELPER_DIR, 'directory');
+                expect(compilation.contextDependencies).to.not.have.members([absFrom]);
+            })
+            .then(done)
+            .catch(done);
+        });
     });
 
     describe('with file in from', () => {
@@ -518,6 +609,23 @@ describe('apply function', () => {
             .catch(done);
         });
 
+        it('can transform target path', (done) => {
+            runEmit({
+                expectedAssetKeys: [
+                    'subdir/test.txt'
+                ],
+                patterns: [{
+                    from: 'file.txt',
+                    transformPath: function(targetPath, absoluteFrom) {
+                        expect(absoluteFrom).to.equal(path.join(HELPER_DIR, 'file.txt'));
+                        return targetPath.replace('file.txt', 'subdir/test.txt');
+                    }
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
         it('warns when file not found', (done) => {
             runEmit({
                 expectedAssetKeys: [],
@@ -526,6 +634,23 @@ describe('apply function', () => {
                 ],
                 patterns: [{
                     from: 'nonexistent.txt'
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
+        it('warns when file not found and stats is undefined', (done) => {
+            runEmit({
+                compiler: new MockCompilerNoStat(),
+                expectedAssetKeys: [],
+                expectedErrors: [
+                    `[copy-webpack-plugin] unable to locate 'nonexistent.txt' at '${HELPER_DIR}${path.sep}nonexistent.txt'`
+                ],
+                patterns: [{
+                    from: 'nonexistent.txt',
+                    to: '.',
+                    toType: 'dir'
                 }]
             })
             .then(done)
@@ -541,6 +666,23 @@ describe('apply function', () => {
                 patterns: [{
                     from: 'file.txt',
                     transform: function() {
+                        throw 'a failure happened';
+                    }
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
+        it('warns when tranformPath failed', (done) => {
+            runEmit({
+                expectedAssetKeys: [],
+                expectedErrors: [
+                    'a failure happened'
+                ],
+                patterns: [{
+                    from: 'file.txt',
+                    transformPath: function() {
                         throw 'a failure happened';
                     }
                 }]
@@ -966,6 +1108,26 @@ describe('apply function', () => {
             .catch(done);
         });
 
+        it('transformPath with promise', (done) => {
+            runEmit({
+                expectedAssetKeys: [
+                    '/some/path/file.txt'
+                ],
+                patterns: [{
+                    from: 'file.txt',
+                    transformPath: function(targetPath, absoluteFrom) {
+                        expect(absoluteFrom).to.have.string(HELPER_DIR);
+
+                        return new Promise((resolve) => {
+                            resolve('/some/path/' + path.basename(targetPath));
+                        });
+                    }
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
         it('same file to multiple targets', (done) => {
             runEmit({
                 expectedAssetKeys: [
@@ -995,6 +1157,25 @@ describe('apply function', () => {
                 ],
                 patterns: [{
                     from: 'directory'
+                }]
+            })
+            .then(done)
+            .catch(done);
+        });
+
+        it('can transform target path of every file in directory', (done) => {
+            runEmit({
+                expectedAssetKeys: [
+                    '/some/path/.dottedfile',
+                    '/some/path/directoryfile.txt',
+                    '/some/path/nestedfile.txt'
+                ],
+                patterns: [{
+                    from: 'directory',
+                    transformPath: function(targetPath, absoluteFrom) {
+                        expect(absoluteFrom).to.have.string(path.join(HELPER_DIR, 'directory'));
+                        return '/some/path/' + path.basename(targetPath);
+                    }
                 }]
             })
             .then(done)
@@ -1206,9 +1387,9 @@ describe('apply function', () => {
                 }]
             })
             .then((compilation) => {
-                const absFrom = path.join(HELPER_DIR, 'directory');
-
-                expect(compilation.contextDependencies).to.have.members([absFrom]);
+                const absFrom = path.resolve(HELPER_DIR, 'directory');
+                const absFromNested = path.resolve(HELPER_DIR, 'directory', 'nested');
+                expect(compilation.contextDependencies).to.have.members([absFrom, absFromNested]);
             })
             .then(done)
             .catch(done);